requirements.txt
config~=0.5.1
selenium~=4.18.1
pymongo~=4.6.1
psycopg2~=2.9.9
python-dotenv~=1.0.1
google-api-python-client